+++ usecase/product/types.go
@@ -2,5 +2,5 @@ package ucproduct
 
 type GetProductParam struct {
 	Category string `json:"category"`
-	Limit    int64  `json:"limit"`
+	Limit    int    `json:"limit"`
 }
